+++ python/mic_eq/config_parts/presets.py
@@ -286,6 +286,7 @@ class Preset:
                 "1.11.4",
                 "1.12.0",
                 "1.12.1",
+                "1.13.0",
             ):
                 if version_tuple < _version_tuple(version):
                     data["version"] = version